% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barRating.R
\name{barRatingClear}
\alias{barRatingClear}
\title{Clear barRating widget}
\usage{
barRatingClear(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{id of the widget}

\item{shiny}{session}
}
\description{
Clear a barRating widget
}

