% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barRating.R
\name{barRatingReadOnly}
\alias{barRatingReadOnly}
\title{Set barRating read-only status}
\usage{
barRatingReadOnly(id, state, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{id of the widget}

\item{state}{logical.  True = set widget to read-only; false = set widget to read/write.}

\item{shiny}{session}
}
\description{
Set barRating read-only status
}

