% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barRating.R
\name{barRatingUpdate}
\alias{barRatingUpdate}
\title{Update a barrating widget with a value}
\usage{
barRatingUpdate(id, value, session = shiny::getDefaultReactiveDomain())
}
\description{
Update a barrating widget with a value
}

