% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barRating.R
\name{barRating}
\alias{barRating}
\title{<Add Title>}
\usage{
barRating(choices, selected = NULL, theme = "bars-1to10", reverse = FALSE,
  includeEmpty = FALSE, width = NULL, height = NULL)
}
\description{
<Add Description>
}

