% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barRating.R
\name{barRatingUpdate}
\alias{barRatingUpdate}
\title{Update barRating widget}
\usage{
barRatingUpdate(id, value, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{id of the widget}

\item{value}{new rating value}

\item{shiny}{session}
}
\description{
Update a barRating widget with a value
}

